+++ Tests/RetentionFixtures/testRetainsOpenClassParameters.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+open class FixtureClass112 {
+    open func doSomething(with value: Int) { }
+}
+
+public class FixtureClass112Retainer {
+    var instance: FixtureClass112?
+
+    public func doSomething() {
+        instance?.doSomething(with: .random(in: 0...10))
+    }
+}