+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - In Swift 5.3 and lower, all optional protocol members are now retained in order to workaround a Swift bug. This bug is resolved in Swift 5.4.
 - Cases of public enums are now also retained when using `--retain-public`.
 - Empty public protocols that have an implementation in extensions are no longer identified as redundant.
+- Foreign protocol method parameters are no longer identified as unused.
 
 ## 2.4.1 (2020-12-20)